Stock Market_FastApi _Classical/requirements.txt
annotated-types
beautifulsoup4
cachetools
certifi
charset-normalizer
contourpy
cycler
debugpy
exceptiongroup
executing
fonttools
frozendict
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
html5lib
httplib2
idna
ipywidgets
joblib
jupyterlab_widgets
kiwisolver
lxml
matplotlib
matplotlib-inline
multitasking
newsapi-python
numpy
pandas
peewee
pillow
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
python-dateutil
pytz
requests
rsa
scikit-learn
scipy
setuptools
soupsieve
threadpoolctl
tqdm
traitlets
tzdata
uritemplate
urllib3
webencodings
widgetsnbextension
yfinance